v0.02 - Updated Neigbourhood size  - Feb 6, 2012 ---------------------------------------
I tested changing the neighbourhood size to range between 1 and 100. 
The following results were found


N size		RSME 
1 			1.4961451171826305
5			1.1831599539083058 
10 			1.134420381930053
20 			1.110594693291348
30			1.1073889468207587
40 			1.1058960051034536
50			1.1045085051581185  << best result
60			1.1053132181268333
70			1.1066042356564234
80			1.1069454038242852
90 			1.1071530413192758
100 		1.107469530198342

From the graph of information above we can see that a neighbourhood size
of 50 gives the best results for our probe data with a fixed size. 

With bit more refinement I found 53 to be the optimal fix size for this
set.

Result RSME: 1.1043132359396504